requirements.txt
groq
python-dotenv
datasets
transformers
flask
numpy
pandas
tqdm
streamlit